requirements.txt
google-api-python-client==1.5.1
httplib2==0.9.2
Jinja2==2.8
oauth2client==2.2.0
pytest==2.9.2
pytest-mock==1.1
webapp2==2.5.2
WebTest==2.0.21
